backend/requirements.txt
# Flask
Flask==3.0.0
flask-cors==4.0.0

# Database
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
psycopg2-binary==2.9.9

# Auth
Flask-JWT-Extended==4.6.0

# Config & env
python-dotenv==1.0.0

# Image handling (for /analyze upload + future OpenCV/TF)
Pillow==10.2.0
Werkzeug==3.0.1
